organize diarization sub-modules and service
